changed the chunk size in the summary interface and deps to int from long

+++ trunk/projects/rlVizLibJava/src/rlVizLib/messaging/environment/EpisodeSummaryChunkResponse.java
@@ -31,9 +31,9 @@ import rlVizLib.messaging.NotAnRLVizMessageException;
 class EpisodeSummaryChunkResponse extends AbstractResponse {
 
     String theLogData;
-    long amountRequested = 0;
+    int amountRequested = 0;
 
-    public EpisodeSummaryChunkResponse(String theLogString, long amountRequested) {
+    public EpisodeSummaryChunkResponse(String theLogString, int amountRequested) {
         this.amountRequested = amountRequested;
         this.theLogData = theLogString;
     }
